Added ability to change the idle emote to follow cursor emote and back on right click
I liked the dance Agnes does when standing still while in cursor following mode, so I grabbed a can of beer and did this.
My testing showed no problems with other functionality, works as it did before.
If you have different plans for right click, I propose changing RMB click to a specific button, but in the end it's up to you.